separate nnet.*
