+++ Makefile
@@ -47,7 +47,7 @@ CXXFLAGS += -std=c++11 -Wextra -Ofast -march=native -mtune=native
 CPPFLAGS += -MMD -MP -Isrc/common -DNDEBUG -DUSE_SSE4
 LDFLAGS  += -llzma -lpthread
 
-TARGETS        += bin/aobaz bin/autousi bin/server bin/gencode bin/playshogi bin/crc64 bin/extract bin/net-test
+TARGETS        += bin/aobaz bin/autousi bin/server bin/playshogi bin/crc64 bin/extract bin/net-test bin/gencode
 AUTOUSI_OBJS   := src/autousi/autousi.o src/autousi/client.o src/autousi/pipe.o src/common/iobase.o src/common/option.o src/common/jqueue.o src/common/xzi.o src/common/err.o src/common/shogibase.o src/common/osi.o
 SERVER_OBJS    := src/server/server.o src/server/listen.o src/server/datakeep.o src/common/iobase.o src/common/xzi.o src/common/jqueue.o src/common/err.o src/common/option.o src/server/logging.o src/common/shogibase.o src/common/osi.o
 GENCODE_OBJS   := src/gencode/gencode.o
@@ -55,7 +55,7 @@ PLAYSHOGI_OBJS := src/playshogi/playshogi.o src/common/option.o src/common/err.o
 CRC64_OBJS     := src/crc64/crc64.o src/common/xzi.o src/common/err.o src/common/iobase.o src/common/osi.o
 EXTRACT_OBJS   := src/extract/extract.o src/common/xzi.o src/common/err.o src/common/iobase.o src/common/osi.o
 OCLDEVS_OBJS   := src/ocldevs/ocldevs.o src/common/err.o
-NET_TEST_OBJS  := src/net-test/net-test.o src/net-test/nnet.o src/common/err.o src/common/iobase.o src/common/shogibase.o src/common/xzi.o src/common/osi.o
+NET_TEST_OBJS  := src/net-test/net-test.o src/net-test/nnet.o src/net-test/nnet-shogi.o src/net-test/nnet-cpu.o src/common/err.o src/common/iobase.o src/common/shogibase.o src/common/xzi.o src/common/osi.o 
 OBJS           := $(AUTOUSI_OBJS) $(SERVER_OBJS) $(GENCODE_OBJS) $(PLAYSHOGI_OBJS) $(CRC64_OBJS) $(EXTRACT_OBJS) $(OCLDEVS_OBJS) $(NET_TEST_OBJS)
 INC_OUT        := src/common/tbl_zkey.inc src/common/tbl_board.inc src/common/tbl_sq.inc src/common/tbl_bmap.inc
 
@@ -99,12 +99,10 @@ clean:
 src/usi-engine/aobaz: FORCE
 	cd src/usi-engine; $(MAKE)
 
-src/autousi/pipe.cpp: bin/gencode
-src/net-test/nnet.cpp : bin/gencode
-src/net-test/net-test.cpp : bin/gencode
-src/server/datakeep.cpp: bin/gencode
-src/common/shogibase.cpp: bin/gencode
-src/playshogi/playshogi.cpp: bin/gencode
+$(AUTOUSI_OBJS:.o=.cpp) : bin/gencode
+$(SERVER_OBJS:.o=.cpp) : bin/gencode
+$(PLAYSHOGI_OBJS:.o=.cpp) : bin/gencode
+$(NET_TEST_OBJS:.o=.cpp) : bin/gencode
 
 -include $(OBJS:.o=.d)
 FORCE:
