+++ src/net-test/nnet-shogi.hpp
@@ -1,0 +1,24 @@
+#pragma once
+#include "shogibase.hpp"
+
+namespace NNAux {
+  unsigned short encode_nnmove(const Action &a, const Color &turn) noexcept;
+}
+
+template<unsigned int Len>
+class NodeNN : public Node<Len> {
+  using uint  = unsigned int;
+  using uchar = unsigned char;
+  struct {
+    BMap bm[Color::ok_size];
+    uchar board[Sq::ok_size];
+    uchar hand[Color::ok_size][Pc::hand_size];
+    uint count_repeat; } _posi[Len];
+  void set_posi() noexcept;
+  
+public:
+  explicit NodeNN() noexcept { set_posi(); }
+  void clear() noexcept { Node<Len>::clear(); set_posi(); }
+  void take_action(const Action &a) noexcept;
+  void encode_input(float *p) const noexcept;
+};
